+++ bytes/src/lib.rs
@@ -4,14 +4,79 @@
 //  Created:
 //    19 Sep 2023, 21:05:57
 //  Last edited:
-//    21 Sep 2023, 13:52:42
+//    22 Sep 2023, 13:00:39
 //  Auto updated?
 //    Yes
 // 
 //  Description:
-//!   A package for easily defining header-like structs that can be parsed
-//!   from bytes with static interpretation.
-// 
+//!   Adds a Rust derive macro that can be used to easily parse header-like bytes that have specified order and interpretation.
+//!   
+//!   # Usage
+//!   The main use-case of this crate is to define structs that parse tightly-packed fields from a string of bytes.
+//!   
+//!   For example, consider implementing a parser for the UDP header using this crate:
+//!   ```rust
+//!   use bytes::{BigEndian, TryFromBytes};
+//!   
+//!   // See the spec from Wikipedia:
+//!   // <https://en.wikipedia.org/wiki/User_Datagram_Protocol#UDP_datagram_structure>
+//!   #[derive(TryFromBytes)]
+//!   struct UdpHeader {
+//!       /// The packet source port.
+//!       #[bytes(dynamic = BigEndian)]
+//!       src_port : u16,
+//!       /// The packet destination port.
+//!       #[bytes(dynamic = BigEndian)]
+//!       dst_port : u16,
+//!       /// The length of the packet, in bytes.
+//!       #[bytes(dynamic = BigEndian)]
+//!       length   : u16,
+//!       /// A checksum for the datagram.
+//!       #[bytes(dynamic = BigEndian)]
+//!       checksum : u16,
+//!   }
+//!   ```
+//!   
+//!   A main feature of the crate is that parser can take runtime information for customization. In the previous example, we used this to force the numbers to be parsed in big-endian byte order; but we can also be more dynamic and refer to previous fields:
+//!   ```rust
+//!   use bytes::TryFromBytes;
+//!   
+//!   #[derive(TryFromBytes)]
+//!   struct Text {
+//!       /// The length of the text we will be parsing
+//!       #[bytes]
+//!       len : usize,   // Note: in bytes
+//!       /// And then the text itself, from UTF-8
+//!       #[bytes(dynamic = len)]
+//!       txt : String,
+//!   }
+//!   
+//!   let input: &[u8] = &[ 13, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x2c, 0x20, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x21 ];
+//!   assert_eq!(Text::try_from_bytes(input).txt, "Hello, world!");
+//!   ```
+//!   
+//!   For more information on the derive macros, refer to the documentation of the crate (and then the `procedural` module). Other examples may be found in the [examples](./bytes/examples/) directory or as examples in the docstrings.
+//!   
+//!   # Installation
+//!   To use this crate into one of your projects, simply add it to your `Cargo.toml` file:
+//!   ```toml
+//!   bytes = { git = "https://github.com/Lut99/bytes-rs" }
+//!   ```
+//!   Optionally, you can commit to a particular tag:
+//!   ```toml
+//!   bytes = { git = "https://github.com/Lut99/bytes-rs", tag = "v1.0.0" }
+//!   ```
+//!   
+//!   To build this crate's documentation and open it, run:
+//!   ```bash
+//!   cargo doc --all-features --no-deps --open
+//!   ```
+//!   in the root of the repository.
+//!   
+//!   ## Features
+//!   The crate has the following features:
+//!   - `derive`: When given, imports the `bytes-derive`-crate which enables the procedural macros (_enabled by default_).
+//  
 
 // Declare the submodules
 pub mod errors;